Create _JC-WH_adaptive_sceneOut
This scenes file is the first 'test' version for the Jubilee Campus to Wollaton Hall walk. This has the adaptive music on the outward leg od the walk and the fixed playlist on the return leg. This inital version has no adaptive functionality other than track 'start' and 'end' points, and one instance of multi-track volume alterations